+++ system/platform-core/src/main/java/org/platformlambda/core/models/AsyncHttpRequest.java
@@ -18,15 +18,19 @@
 
 package org.platformlambda.core.models;
 
+import io.netty.handler.codec.http.cookie.ServerCookieEncoder;
 import org.platformlambda.core.serializers.PayloadMapper;
 import org.platformlambda.core.serializers.SimpleMapper;
 import org.platformlambda.core.util.Utility;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.*;
 
 public class AsyncHttpRequest {
+    private static final Logger log = LoggerFactory.getLogger(AsyncHttpRequest.class);
     private static final String HTTP_HEADERS = "headers";
     private static final String HTTP_METHOD = "method";
     private static final String IP_ADDRESS = "ip";
@@ -252,13 +256,17 @@ public class AsyncHttpRequest {
     }
 
     public String getCookie(String key) {
-        String value = caseInsensitiveGet(cookies, key);
-        return value.replace("\r", "").replace("\n", "");
+        return caseInsensitiveGet(cookies, key);
     }
 
     public AsyncHttpRequest setCookie(String key, String value) {
         // avoid CR/LF attack
-        setNonNullValue(cookies, key, value.replace("\r", "").replace("\n", ""));
+        try {
+            ServerCookieEncoder.STRICT.encode(key, value);
+            cookies.put(key, value);
+        } catch(Exception e) {
+            log.warn("Invalid cookie {} ignored - {}", key, e.getMessage());
+        }
         return this;
     }