+++ system/platform-core/src/main/java/org/platformlambda/core/models/EventEnvelope.java
@@ -486,7 +486,6 @@ public class EventEnvelope {
                 } else {
                     this.headers.put(key, v);
                 }
-
             } else {
                 this.headers.put(key, v);
             }