+++ CHANGELOG.md
@@ -8,6 +8,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## Version 4.2.12, 2/12/2025
+
+### Added
+
+N/A
+
+### Removed
+
+N/A
+
+### Changed
+
+1. Use ServerCookieEncoder.STRICT.encode() method to detect invalid cookie value
+2. Update vertx to 4.5.13 and Netty to 4.1.118.Final to address security vulnerabilities
+
+---
 ## Version 4.2.11, 2/11/2025
 
 ### Added